veles_mainwindow.h/cc split into two files.
